+++ src/pmarlo/data/demux_dataset.py
@@ -16,8 +16,8 @@ from typing import Any, Callable, Iterable, List, Optional, Sequence
 
 import numpy as np
 
-from pmarlo.utils.errors import TemperatureConsistencyError
 from pmarlo.features.pairs import scaled_time_pairs
+from pmarlo.utils.errors import TemperatureConsistencyError
 
 
 @dataclass(frozen=True)